Accelerate cosimulation IO (#434)
Added a feature that slows the manycore "core_clk" during large IO operations. This means that IO proceeds much faster in realtime because there are less toggles in the VCS activity list. 

The clock slowdown is controlled by a "Virtual DIP switch" that is controlled by the AWS API. 

There were several issues during simulation. First, we had to instantiate a clock mux to avoid spurious assertions while switching between clocks. Second, we had to make sure that "asynchronous" resets were actually asserted at the same time. 

Credit to @leonardxiang for debugging, and @tommydcjung for implementing the initial version of this.